feat: Enable gameplay functionality
Introduced the ability for users to input a puzzle and play the game.
Users can input clues and leave blank spaces by entering 0, and they
can begin the puzzle by pressing 000. Additionally, a new function,
chkwin(), is added to check whether the game has been won or not. The
game will alert the user once the puzzle is solved. To exit the game
during gameplay, users can enter 000.