feat: Add spreadsheet for input assistance
Introduced a spreadsheet feature to assist users in inputting clues.
As entering clues by referencing cell addresses can be challenging, the
spreadsheet provides a convenient solution. Users can directly input
values into the grid using the spreadsheet interface. The spreadsheet
generates a corresponding string of numbers, which can be copied and
pasted into the console when prompted for puzzle input. To leave a cell
blank in the puzzle, users can enter 0 in the corresponding cell within
the spreadsheet.