fix: Prevent user from editing puzzle input
This commit addresses a bug that allowed users to edit clues by entering
the address of a clue in-game. To resolve this issue, a separate 9x9
array is introduced to store the initial clues. The main array is
refreshed using the respuz() function whenever the user inputs a value.
As a result, users are no longer able to edit the initial clues during
the game.