feat: Add game menu and solver
This commit introduces an extensive game menu, offering users the choice
to play the puzzle or find solution to any Sudoku puzzle. Although
the help and about sections are currently under construction, the game
option presents users with a list of difficulty levels to select from.
Additionally, the solver mode provides a submenu for various input
options, allowing users to solve the puzzle and exit.