fix: Fix minor bugs
This commit addresses a minor bug that caused the program to crash at
times due to the presence of garbage values in the arrays. Additionally,
a bug in the solver submenu that caused unexpected behavior has been
resolved. Furthermore, winning the puzzle no longer exits the program.
Some minor UI tweaks have also been included in this fix.