feat: Allow using arrow keys in grid
This commit introduces the ability for the user to navigate through
the grid using arrow keys. Previously, the user had to type address and
press the enter key after each input, which was inconvenient. With this
update, pressing arrow keys allows for smooth movement within the grid.
As a result, the functions allinput(), edtinput() and the spreadsheet
gridGen.xlsx are no longer necessary and have been removed.

Additionally, the key 'q' is now designated as the global key for
displaying the menu or help, providing a more streamlined user
experience.