fix: Generate more randomized puzzle
The previous method for generating puzzles had limitations as it
resulted in repeated grids due to the use of backtracking in the solving
algorithm. The algorithm randomly filled 5 elements and then solved the
grid, which led to collisions and repetitive patterns. To address this,
a new approach is implemented.

The diagonal blocks (subgrids) of the grid are filled with values 1-9,
as they are independent (there are other independent arrangement as
well) and do not collide with each other. The solve() function is then
used to generate the remaining part of the grid, resulting in a more
randomized puzzle generation.